Eliminated cells in favour of a list (for now)
Cells existed to allow the first argument of each operation to take
a reference to a register or memory location so it could be set.

This has been replaced by first observing that the first argument
to each operation is always a reference (register or memory) and never
a literal and that the second argument can always be deferences before
being passed in and so can be treated as a literal value in all cases.

The remaining trick, then, was to treat registers and memory locations
with the same addressing scheme by placing the registers in offsets
0x10000+ in the "memory" list.